requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
numpy>=1.24.0
torch>=2.0.0
opencv-python>=4.8.0
mediapipe==0.10.9
pillow>=10.0.0
